Update patch set 1
Patch Set 1:

I had a look into the build errors.  Some of them are trivial to fix (just changed packages for some classes), but others are not.  For example the API to manipulate project watches has significantly changed.

Marco, do you have a work in progress change for this?

Patch-set: 1