Add Gatling e2e-test for rename-project
A successful test will guarantee all the steps as performed. This tests
the plugin feature through creating a project, then renaming it, and
then deleting the renamed project. If the rename fails, then only the
project gets created. If the rename goes through but gives unexpected
results, then the project deletion step will fail due to not finding
the renamed project (to delete).

Add a README file that introduces how to start using this new test.

The test project is renamed to a unique new name, _RENAMED in the json,
automatically set by the test implementation. This is in line with how
such a use of an underscore prefix was documented in [1]. The original
project name is also automatically set that way. Both names are based on
the name of the test class, so named after "RenameProject".

[1] https://gerrit-documentation.storage.googleapis.com/Documentation/3.0.12/dev-e2e-tests.html#_environment_properties

Change-Id: If086299127b8eb418c693e07ce221dd29cc5468d
(cherry picked from commit 98e1df4b5d7a89e67dd75fee54c79d6dc9fc2d10)