Introduce an optional configuration to define a regex for naming rules
Before this change, the plugin made checks on the name to verify that
the new project name is not null and that it doesn't exist. Therefore,
the new project name can have aname with any set of characters while
complying to the naming rules. However, due to IRI (not URI) and
backward compatibility, the non-ASCII characters may be represented in a
way that can cause issues [1].

This change introduces a configurable parameter that can be set to
define a regex that is then enforced upon renaming a project to ensure
backward compatibility.

[1] https://en.wikipedia.org/wiki/Internationalized_Resource_Identifier

Bug: Issue 14297
Change-Id: Ia049cb399afe1091626cafebfb3d64f72f41e61f