Ensure project cache state during revert of rename
Since I3724dfead the project cache caches non-existing projects. During
the revert of a rename, this caused indexing of changes to fail, since
the cache entry of the old project was outdated.

The cache entry of the old project is now evicted as part of the cache
revert. This removes the entry that was created, when the project was
renamed.

Change-Id: I3abc0f1669949324cfc1cc4ea7322e230a556779