Fix project state not reverting to ACTIVE after rename operation
In change Ie1694eb, logic was added to update the project state to
READ_ONLY before starting the rename operation. However, the project
state was not being reverted to ACTIVE after the rename operation on
replica nodes.

This change moves the logic to unlock the project outside the replica
check, ensuring the project state is correctly reset to ACTIVE after the
rename completes.

Change-Id: Idbadff40543ce8bf7a512839b60ca02a5365e5c8