Notify replication plugin about the deletion
When used with replicas, the project under the new name is considered a
project creation and is replicated out of the box by the replication
plugin.

The project under the old name is not deleted by the replication plugin
which is surprising since a deletion made by the `delete-project` plugin
is replicated (if the replica has `replicateProjectDeletions=true`).

The reason it does not work is rename-project has its own system to
replicate the deletion by issueing the rename-project command to the
Gerrit daemons on the replicas.

This causes several issues:
* logic is duplicated
* replicas must have a Gerrit daemon running (I have a use case of a
  spare replicas that receives replications over OpenSSH but does not
  have daemon running)
* the plugin requires the target replicas to be redundantly configured
  under `[plugin "rename-project"]` which requires keeping the list in
  sync with the replication config
* It requires a privileged user on the replicas Gerrit
* When using ssh, a ssh key pair and trusted ssh host keys must be
  configured in order to interact with the Gerrit daemons (the
  replication plugin uses non Gerrit users over eg OpenSSH).

Since the project under the new name is created by the replication
plugin, I thought it could also handle the deletion. The replication
plugin implements ProjectDeletedListener and that is how a project is
deleted on replicas when it is deleted by the delete-project plugin.

Copy from delete-project RepositoryDelete the logic to send a
ProjectDeletedListener.Event() on deletion.

Change-Id: Id720329fcfdb85fc4b1b4d8dc35ed26bbacc3fe7